+++ src/lib.rs
@@ -1,5 +1,6 @@
 mod parser;
 pub use parser::parse_status::ParseStatus;
+pub use parser::relation::Area;
 pub use parser::relation::PublicTransport;
 pub use parser::relation::Relation;
 pub use parser::Parser;