+++ tests/integration_test.rs
@@ -2,7 +2,7 @@ use osmptparser::Parser;
 
 #[test]
 fn get_public_transports() {
-    let parser = Parser::new("tests/test.pbf", 1);
+    let parser = Parser::new_ptv2("tests/test.pbf", 1);
     let ptsvec = parser.get_public_transports(1500_f64);
     let mut pts = ptsvec.iter().collect::<Vec<_>>();
     pts.sort_by(|a, b| a.id.cmp(&b.id));